Personal Training/Security/Script/requirements.txt
pandas
python-dotenv
snowflake-connector-python[secure-local-storage,pandas]
email
smtplib
base64